Refactor: 컴포넌트 관심사 분리를 위한 MyPage 초기 구조 변경
